Refactor to avoid merge conflicts
