+++ smartcon_app/lib/controller/auth.dart
@@ -1,6 +1,6 @@
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:google_sign_in/google_sign_in.dart';
-import 'package:smartcon_app/models/user.dart';
+import 'package:smartcon_app/model/user.dart';
 
 import 'database.dart';
