+++ smartcon_app/lib/main.dart
@@ -2,8 +2,8 @@ import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
 import 'package:provider/provider.dart';
-import 'package:smartcon_app/services/auth.dart';
-import 'package:smartcon_app/models/user.dart';
+import 'package:smartcon_app/controller/auth.dart';
+import 'package:smartcon_app/model/user.dart';
 import 'package:smartcon_app/wrapper.dart';
 
 void main() async {
