+++ smartcon_app/lib/model/conference.dart
