+++ smartcon_app/lib/model/session.dart
