+++ smartcon_app/lib/view/common/sessionTile.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
-import 'package:smartcon_app/models/session.dart';
+import 'package:smartcon_app/model/session.dart';
 import 'package:url_launcher/url_launcher.dart';
 
 class SessionTile extends StatelessWidget {
