+++ smartcon_app/lib/view/sessionSuggestions/answerQuiz.dart
@@ -2,10 +2,10 @@ import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:grouped_buttons/grouped_buttons.dart';
 import 'package:provider/provider.dart';
-import 'package:smartcon_app/models/session.dart';
-import 'package:smartcon_app/models/user.dart';
-import 'package:smartcon_app/screens/sessionSuggestions/sessionSuggestions.dart';
-import 'package:smartcon_app/services/database.dart';
+import 'package:smartcon_app/model/session.dart';
+import 'package:smartcon_app/model/user.dart';
+import 'package:smartcon_app/view/sessionSuggestions/sessionSuggestions.dart';
+import 'package:smartcon_app/controller/database.dart';
 
 class AnswerQuiz extends StatefulWidget {
