+++ smartcon_app/lib/view/sessionSuggestions/sessionSuggestionsCode.dart
@@ -1,11 +1,10 @@
 import 'dart:ui';
-
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-import 'package:smartcon_app/models/conference.dart';
-import 'package:smartcon_app/models/user.dart';
-import 'package:smartcon_app/screens/sessionSuggestions/sessionSuggestions.dart';
-import 'package:smartcon_app/services/database.dart';
+import 'package:smartcon_app/model/conference.dart';
+import 'package:smartcon_app/model/user.dart';
+import 'package:smartcon_app/view/sessionSuggestions/sessionSuggestions.dart';
+import 'package:smartcon_app/controller/database.dart';
 
 import 'answerQuiz.dart';
