+++ smartcon_app/lib/wrapper.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-import 'package:smartcon_app/models/user.dart';
-import 'package:smartcon_app/screens/homePage.dart';
-import 'package:smartcon_app/screens/login.dart';
+import 'package:smartcon_app/model/user.dart';
+import 'package:smartcon_app/view/homePage.dart';
+import 'package:smartcon_app/view/login.dart';
 
 class Wrapper extends StatelessWidget {
   @override
